expansions
The function definition has been expanded to accept an additional parameter, "maxWait", to set the maximum wait time.

The validation of the function parameters has been expanded to include checks for "maxWait".

The logic for executing the debounced function has been reworked to include the maximum wait time.

A new flag, "canceled", has been added to keep track of whether the debounced function has been canceled.

The "cancel" method has been updated to clear the timeout and set the canceled flag.

The "result" method has been updated to return the results of the debounced function.